fix: iOS LaTeX rendering by extracting shared markdown parser to commonMain
