+++ feature/chat/src/androidMain/kotlin/com/garfiec/librechat/feature/chat/components/MarkdownContent.kt
@@ -206,7 +206,7 @@ actual fun MarkdownContent(
                         occurrenceOffset += segmentOccurrences
                     } else {
                         val hasCitations = remember(segment.text) {
-                            segment.text.contains(CITATION_REGEX)
+                            segment.text.contains(CITATION_DETECT_REGEX)
                         }
                         if (hasCitations) {
                             CitationText(
@@ -326,420 +326,6 @@ private fun MarkdownTextSegment(
     }
 }
 
-internal sealed interface MarkdownSegment {
-    data class TextBlock(val text: String) : MarkdownSegment
-    data class CodeBlock(val code: String, val language: String?) : MarkdownSegment
-    data class LatexBlock(val latex: String) : MarkdownSegment
-    data class InlineLatexText(val segments: List<InlineSegment>) : MarkdownSegment
-    data class Table(
-        val headers: List<String>,
-        val alignments: List<TableCellAlignment>,
-        val rows: List<List<String>>,
-    ) : MarkdownSegment
-}
-
-/** Column alignment parsed from the separator row of a markdown table. */
-internal enum class TableCellAlignment {
-    LEFT, CENTER, RIGHT,
-}
-
-internal sealed interface InlineSegment {
-    data class Text(val text: String) : InlineSegment
-    data class Latex(val latex: String) : InlineSegment
-}
-
-internal sealed interface InlineSpan {
-    val text: String
-
-    data class Plain(override val text: String) : InlineSpan
-    data class Bold(override val text: String) : InlineSpan
-    data class Italic(override val text: String) : InlineSpan
-    data class InlineCode(override val text: String) : InlineSpan
-}
-
-// Pre-compiled regex patterns for markdown/LaTeX parsing (avoid recreating on every call)
-private val CODE_BLOCK_REGEX = Regex("```(\\w*)[^\\S\\n]*\\n([\\s\\S]*?)```")
-private val BLOCK_LATEX_REGEX = Regex("\\$\\$([\\s\\S]+?)\\$\\$|\\\\\\[([\\s\\S]+?)\\\\\\]")
-private val INLINE_LATEX_REGEX = Regex("(?<!\\$)\\$(?!\\$)(.+?)(?<!\\$)\\$(?!\\$)|\\\\\\((.+?)\\\\\\)")
-private val INLINE_MARKDOWN_REGEX = Regex("`([^`]+)`|\\*\\*(.+?)\\*\\*|\\*(.+?)\\*")
-private val CITATION_REGEX = Regex("""\[\d+]|\u3010\d+\u2020""")
-
-/**
- * Matches a GFM-style markdown table separator row. The separator must contain only
- * pipes, dashes, colons (for alignment), and whitespace.
- */
-private val TABLE_SEPARATOR_REGEX = Regex("^\\|?\\s*:?-{1,}:?\\s*(\\|\\s*:?-{1,}:?\\s*)*\\|?$")
-
-/**
- * HTML tag names that indicate a raw HTML block when found at the start of a TextBlock.
- * These are block-level elements that LLMs commonly produce when generating web pages.
- */
-private val HTML_BLOCK_TAG_NAMES = setOf(
-    "html", "head", "body", "div", "section", "article", "main", "header", "footer",
-    "nav", "aside", "table", "form", "fieldset", "details", "dialog", "figure",
-    "figcaption", "template", "canvas", "svg", "video", "audio", "iframe",
-    "p", "ul", "ol", "li", "dl", "dt", "dd", "pre", "blockquote", "hr",
-)
-
-/**
- * Matches an opening HTML tag at the start of a string (after optional whitespace).
- * Group 1 captures the tag name. Case-insensitive.
- */
-private val HTML_OPENING_TAG_REGEX = Regex(
-    "^\\s*<(!DOCTYPE\\s+html|\\w+)",
-    RegexOption.IGNORE_CASE,
-)
-
-/** Characters/keywords that indicate LaTeX content rather than a dollar amount. */
-private val LATEX_INDICATORS = setOf('\\', '^', '_', '{', '}')
-private val LATEX_KEYWORDS = listOf("frac", "sqrt", "sum", "int", "lim", "infty", "alpha", "beta")
-
-/** Returns true if the content looks like LaTeX rather than a dollar amount. */
-private fun looksLikeLatex(content: String): Boolean {
-    if (content.any { it in LATEX_INDICATORS }) return true
-    return LATEX_KEYWORDS.any { keyword -> content.contains(keyword) }
-}
-
-/**
- * Returns true if the text appears to be a raw HTML block that should be rendered
- * as a code block rather than passed to the markdown renderer (which would garble it).
- *
- * Heuristic: starts with an HTML tag whose name is in [HTML_BLOCK_TAG_NAMES] or is
- * a DOCTYPE declaration, AND contains at least one closing tag or is self-contained.
- */
-private fun looksLikeHtmlBlock(text: String): Boolean {
-    val match = HTML_OPENING_TAG_REGEX.find(text) ?: return false
-    val tagName = match.groupValues[1].lowercase().let {
-        // Normalize "!doctype html" to "html"
-        if (it.startsWith("!doctype")) "html" else it
-    }
-    if (tagName !in HTML_BLOCK_TAG_NAMES) return false
-    // Require a closing tag or self-closing indicator to avoid false positives
-    // on lines that just happen to start with "<p" etc.
-    val hasClosingTag = text.contains("</$tagName", ignoreCase = true) ||
-        text.contains("/>")
-    return hasClosingTag
-}
-
-/**
- * Scans TextBlock segments for raw HTML blocks and converts them to CodeBlock
- * segments with language="html". This prevents the mikepenz markdown renderer
- * from silently dropping HTML tags.
- *
- * A TextBlock is treated as a raw HTML block if [looksLikeHtmlBlock] returns true.
- * Mixed content (text + HTML) is split: lines before the HTML become a TextBlock,
- * the HTML block becomes a CodeBlock, and lines after become a TextBlock.
- */
-private fun extractHtmlBlocks(segments: List<MarkdownSegment>): List<MarkdownSegment> {
-    val result = mutableListOf<MarkdownSegment>()
-
-    for (segment in segments) {
-        if (segment !is MarkdownSegment.TextBlock) {
-            result.add(segment)
-            continue
-        }
-
-        val text = segment.text
-        if (looksLikeHtmlBlock(text)) {
-            // Entire block is HTML -- render as a code block
-            result.add(MarkdownSegment.CodeBlock(text.trim(), "html"))
-            continue
-        }
-
-        // Check if the block contains an embedded HTML section starting on its own line.
-        // Split into lines and find the first line that opens an HTML block.
-        val lines = text.split('\n')
-        var htmlStart = -1
-        for (i in lines.indices) {
-            val lineText = lines.subList(i, lines.size).joinToString("\n")
-            if (looksLikeHtmlBlock(lineText)) {
-                htmlStart = i
-                break
-            }
-        }
-
-        if (htmlStart < 0) {
-            // No HTML found; keep as-is
-            result.add(segment)
-            continue
-        }
-
-        // Flush preceding text lines
-        if (htmlStart > 0) {
-            val preceding = lines.subList(0, htmlStart).joinToString("\n").trim()
-            if (preceding.isNotEmpty()) {
-                result.add(MarkdownSegment.TextBlock(preceding))
-            }
-        }
-
-        // The rest from htmlStart onward is the HTML block
-        val htmlContent = lines.subList(htmlStart, lines.size).joinToString("\n").trim()
-        result.add(MarkdownSegment.CodeBlock(htmlContent, "html"))
-    }
-
-    return result
-}
-
-/**
- * Splits raw text into code-block, LaTeX-block, table, inline-LaTeX, and text-block segments.
- *
- * Parsing order:
- * 1. Extract fenced code blocks (``` ... ```)
- * 2. Detect raw HTML blocks in remaining TextBlocks and convert to CodeBlocks
- * 3. Within non-code segments, extract block LaTeX (`$$...$$`)
- * 4. Within remaining text blocks, extract GFM-style tables
- * 5. Within remaining text, detect inline LaTeX (`$...$`) and split accordingly
- */
-internal fun parseMarkdownSegments(text: String): List<MarkdownSegment> {
-    // --- Pass 1: split on fenced code blocks ---
-    val afterCodeBlocks = mutableListOf<MarkdownSegment>()
-    var lastIndex = 0
-
-    CODE_BLOCK_REGEX.findAll(text).forEach { match ->
-        if (match.range.first > lastIndex) {
-            val textBefore = text.substring(lastIndex, match.range.first).trim()
-            if (textBefore.isNotEmpty()) {
-                afterCodeBlocks.add(MarkdownSegment.TextBlock(textBefore))
-            }
-        }
-        val language = match.groupValues[1].ifEmpty { null }
-        val code = match.groupValues[2].trimEnd()
-        val langLower = language?.lowercase()
-        if (langLower == "latex" || langLower == "tex" || langLower == "math") {
-            // Treat latex/tex/math code blocks as text so LaTeX extraction
-            // passes handle $...$ and $$...$$ delimiters properly
-            afterCodeBlocks.add(MarkdownSegment.TextBlock(code))
-        } else {
-            afterCodeBlocks.add(MarkdownSegment.CodeBlock(code, language))
-        }
-        lastIndex = match.range.last + 1
-    }
-
-    if (lastIndex < text.length) {
-        val remaining = text.substring(lastIndex).trim()
-        if (remaining.isNotEmpty()) {
-            afterCodeBlocks.add(MarkdownSegment.TextBlock(remaining))
-        }
-    }
-
-    if (afterCodeBlocks.isEmpty() && text.isNotBlank()) {
-        afterCodeBlocks.add(MarkdownSegment.TextBlock(text))
-    }
-
-    // --- Pass 2: detect raw HTML blocks and convert to CodeBlocks ---
-    val afterHtmlBlocks = extractHtmlBlocks(afterCodeBlocks)
-
-    // --- Pass 3: split TextBlocks on block LaTeX ($$...$$ or \[...\]) ---
-    val afterBlockLatex = mutableListOf<MarkdownSegment>()
-
-    for (segment in afterHtmlBlocks) {
-        if (segment !is MarkdownSegment.TextBlock) {
-            afterBlockLatex.add(segment)
-            continue
-        }
-        var segLastIndex = 0
-        val segText = segment.text
-
-        BLOCK_LATEX_REGEX.findAll(segText).forEach { match ->
-            if (match.range.first > segLastIndex) {
-                val before = segText.substring(segLastIndex, match.range.first).trim()
-                if (before.isNotEmpty()) {
-                    afterBlockLatex.add(MarkdownSegment.TextBlock(before))
-                }
-            }
-            val latexContent = (match.groupValues[1].ifEmpty { match.groupValues[2] }).trim()
-            if (latexContent.isNotEmpty()) {
-                afterBlockLatex.add(MarkdownSegment.LatexBlock(latexContent))
-            }
-            segLastIndex = match.range.last + 1
-        }
-
-        if (segLastIndex < segText.length) {
-            val remaining = segText.substring(segLastIndex).trim()
-            if (remaining.isNotEmpty()) {
-                afterBlockLatex.add(MarkdownSegment.TextBlock(remaining))
-            }
-        } else if (segLastIndex == 0) {
-            // No block LaTeX found; keep original segment
-            afterBlockLatex.add(segment)
-        }
-    }
-
-    // --- Pass 4: extract markdown tables from TextBlocks ---
-    val afterTables = mutableListOf<MarkdownSegment>()
-
-    for (segment in afterBlockLatex) {
-        if (segment !is MarkdownSegment.TextBlock) {
-            afterTables.add(segment)
-            continue
-        }
-        afterTables.addAll(extractTables(segment.text))
-    }
-
-    // --- Pass 5: detect inline LaTeX ($...$ or \(...\)) within remaining TextBlocks ---
-    val finalSegments = mutableListOf<MarkdownSegment>()
-
-    for (segment in afterTables) {
-        if (segment !is MarkdownSegment.TextBlock) {
-            finalSegments.add(segment)
-            continue
-        }
-
-        val matches = INLINE_LATEX_REGEX.findAll(segment.text)
-            .filter { match ->
-                // \(...\) is always LaTeX; $...$ needs heuristic check
-                val dollarContent = match.groupValues[1]
-                val parenContent = match.groupValues[2]
-                if (parenContent.isNotBlank()) {
-                    true
-                } else {
-                    dollarContent.isNotBlank() && looksLikeLatex(dollarContent)
-                }
-            }
-            .toList()
-
-        if (matches.isEmpty()) {
-            finalSegments.add(segment)
-            continue
-        }
-
-        // Build mixed inline segments
-        val inlineSegments = mutableListOf<InlineSegment>()
-        var segLastIndex = 0
-        val segText = segment.text
-
-        for (match in matches) {
-            if (match.range.first > segLastIndex) {
-                val before = segText.substring(segLastIndex, match.range.first)
-                if (before.isNotEmpty()) {
-                    inlineSegments.add(InlineSegment.Text(before))
-                }
-            }
-            val content = (match.groupValues[1].ifEmpty { match.groupValues[2] }).trim()
-            inlineSegments.add(InlineSegment.Latex(content))
-            segLastIndex = match.range.last + 1
-        }
-
-        if (segLastIndex < segText.length) {
-            val remaining = segText.substring(segLastIndex)
-            if (remaining.isNotEmpty()) {
-                inlineSegments.add(InlineSegment.Text(remaining))
-            }
-        }
-
-        finalSegments.add(MarkdownSegment.InlineLatexText(inlineSegments))
-    }
-
-    return finalSegments
-}
-
-/**
- * Extracts markdown tables from a text block, splitting it into alternating
- * TextBlock and Table segments. A valid table requires:
- * - A header row with pipe-delimited cells
- * - A separator row matching [TABLE_SEPARATOR_REGEX]
- * - At least one data row
- */
-private fun extractTables(text: String): List<MarkdownSegment> {
-    val lines = text.split('\n')
-    val result = mutableListOf<MarkdownSegment>()
-    val buffer = mutableListOf<String>()
-    var i = 0
-
-    while (i < lines.size) {
-        // Check if lines[i] could be a table header row and lines[i+1] a separator
-        if (i + 2 < lines.size && isTableRow(lines[i]) && isTableSeparator(lines[i + 1])) {
-            // Flush any buffered text before the table
-            if (buffer.isNotEmpty()) {
-                val preceding = buffer.joinToString("\n").trim()
-                if (preceding.isNotEmpty()) {
-                    result.add(MarkdownSegment.TextBlock(preceding))
-                }
-                buffer.clear()
-            }
-
-            val headerCells = parseTableRow(lines[i])
-            val alignments = parseAlignments(lines[i + 1], headerCells.size)
-            val dataRows = mutableListOf<List<String>>()
-            var j = i + 2
-
-            while (j < lines.size && isTableRow(lines[j])) {
-                val rowCells = parseTableRow(lines[j])
-                // Pad or truncate to match header column count
-                val normalized = List(headerCells.size) { col ->
-                    rowCells.getOrElse(col) { "" }
-                }
-                dataRows.add(normalized)
-                j++
-            }
-
-            if (dataRows.isNotEmpty()) {
-                result.add(MarkdownSegment.Table(headerCells, alignments, dataRows))
-                i = j
-            } else {
-                // Not a valid table (no data rows); treat header+separator as text
-                buffer.add(lines[i])
-                buffer.add(lines[i + 1])
-                i += 2
-            }
-        } else {
-            buffer.add(lines[i])
-            i++
-        }
-    }
-
-    // Flush remaining buffered text
-    if (buffer.isNotEmpty()) {
-        val remaining = buffer.joinToString("\n").trim()
-        if (remaining.isNotEmpty()) {
-            result.add(MarkdownSegment.TextBlock(remaining))
-        }
-    }
-
-    return result
-}
-
-/**
- * Returns true if the line looks like a pipe-delimited table row.
- * Requires either leading/trailing pipes or at least two pipe characters,
- * to avoid false positives on prose that contains a single `|`.
- */
-private fun isTableRow(line: String): Boolean {
-    val trimmed = line.trim()
-    if (trimmed.isEmpty()) return false
-    return trimmed.startsWith('|') || trimmed.endsWith('|') || trimmed.count { it == '|' } >= 2
-}
-
-/** Returns true if the line matches a GFM table separator pattern. */
-private fun isTableSeparator(line: String): Boolean {
-    return TABLE_SEPARATOR_REGEX.matches(line.trim())
-}
-
-/** Splits a pipe-delimited table row into trimmed cell values. */
-private fun parseTableRow(line: String): List<String> {
-    val trimmed = line.trim()
-    // Remove leading and trailing pipes if present
-    val inner = trimmed.removePrefix("|").removeSuffix("|")
-    return inner.split('|').map { it.trim() }
-}
-
-/**
- * Parses alignment indicators from a separator row.
- * `:---` = LEFT, `:---:` = CENTER, `---:` = RIGHT, `---` = LEFT (default)
- */
-private fun parseAlignments(separatorLine: String, columnCount: Int): List<TableCellAlignment> {
-    val cells = parseTableRow(separatorLine)
-    return List(columnCount) { col ->
-        val cell = cells.getOrElse(col) { "---" }.trim()
-        val startsColon = cell.startsWith(':')
-        val endsColon = cell.endsWith(':')
-        when {
-            startsColon && endsColon -> TableCellAlignment.CENTER
-            endsColon -> TableCellAlignment.RIGHT
-            else -> TableCellAlignment.LEFT
-        }
-    }
-}
 
 /**
  * Wraps a [MarkdownTable] in a [Box] with a fullscreen expand button overlaid on
@@ -1005,32 +591,3 @@ private fun TableCell(
     }
 }
 
-/**
- * Parses inline markdown (bold, italic, inline code) within a text block.
- * Calls [onSpan] for each discovered span.
- */
-internal fun parseInlineMarkdown(text: String, onSpan: (InlineSpan) -> Unit) {
-    var lastIndex = 0
-
-    INLINE_MARKDOWN_REGEX.findAll(text).forEach { match ->
-        if (match.range.first > lastIndex) {
-            onSpan(InlineSpan.Plain(text.substring(lastIndex, match.range.first)))
-        }
-        when {
-            match.groupValues[1].isNotEmpty() -> {
-                onSpan(InlineSpan.InlineCode(match.groupValues[1]))
-            }
-            match.groupValues[2].isNotEmpty() -> {
-                onSpan(InlineSpan.Bold(match.groupValues[2]))
-            }
-            match.groupValues[3].isNotEmpty() -> {
-                onSpan(InlineSpan.Italic(match.groupValues[3]))
-            }
-        }
-        lastIndex = match.range.last + 1
-    }
-
-    if (lastIndex < text.length) {
-        onSpan(InlineSpan.Plain(text.substring(lastIndex)))
-    }
-}
